refactor: reuse chat attachment widgets in chat list renderer
