+++ src/vs/workbench/contrib/chat/browser/chatAttachmentWidgets.ts
@@ -17,9 +17,9 @@ import { URI } from '../../../../base/common/uri.js';
 import { localize } from '../../../../nls.js';
 import { ICommandService } from '../../../../platform/commands/common/commands.js';
 import { ITextEditorOptions } from '../../../../platform/editor/common/editor.js';
-import { FileKind } from '../../../../platform/files/common/files.js';
+import { FileKind, IFileService } from '../../../../platform/files/common/files.js';
 import { IHoverService } from '../../../../platform/hover/browser/hover.js';
-import { IInstantiationService } from '../../../../platform/instantiation/common/instantiation.js';
+import { IInstantiationService, ServicesAccessor } from '../../../../platform/instantiation/common/instantiation.js';
 import { ILabelService } from '../../../../platform/label/common/label.js';
 import { IOpenerService, OpenInternalOptions } from '../../../../platform/opener/common/opener.js';
 import { IThemeService, FolderThemeIcon } from '../../../../platform/theme/common/themeService.js';
@@ -27,16 +27,30 @@ import { IResourceLabel, ResourceLabels, IFileLabelOptions } from '../../../brow
 import { revealInSideBarCommand } from '../../files/browser/fileActions.contribution.js';
 import { IChatRequestPasteVariableEntry, IChatRequestVariableEntry, IElementVariableEntry, INotebookOutputVariableEntry, OmittedState } from '../common/chatModel.js';
 import { ILanguageModelChatMetadataAndIdentifier, ILanguageModelsService } from '../common/languageModels.js';
-import { hookUpResourceAttachmentDragAndContextMenu, hookUpSymbolAttachmentDragAndContextMenu } from './chatContentParts/chatAttachmentsContentPart.js';
+import { chatAttachmentResourceContextKey } from './chatContentParts/chatAttachmentsContentPart.js';
 import { KeyCode } from '../../../../base/common/keyCodes.js';
 import { basename, dirname } from '../../../../base/common/path.js';
-import { IContextKeyService } from '../../../../platform/contextkey/common/contextkey.js';
-import { MenuId } from '../../../../platform/actions/common/actions.js';
+import { IContextKey, IContextKeyService, IScopedContextKeyService } from '../../../../platform/contextkey/common/contextkey.js';
+import { IMenuService, MenuId } from '../../../../platform/actions/common/actions.js';
 import { ITelemetryService } from '../../../../platform/telemetry/common/telemetry.js';
 import { INotebookService } from '../../notebook/common/notebookService.js';
 import { CellUri } from '../../notebook/common/notebookCommon.js';
 import { ThemeIcon } from '../../../../base/common/themables.js';
 import { IEditorService } from '../../../services/editor/common/editorService.js';
+import { EditorContextKeys } from '../../../../editor/common/editorContextKeys.js';
+import { LanguageFeatureRegistry } from '../../../../editor/common/languageFeatureRegistry.js';
+import { ILanguageFeaturesService } from '../../../../editor/common/services/languageFeatures.js';
+import { ITextModelService } from '../../../../editor/common/services/resolverService.js';
+import { fillInSymbolsDragData } from '../../../../platform/dnd/browser/dnd.js';
+import { fillEditorsDragData } from '../../../browser/dnd.js';
+import { StandardMouseEvent } from '../../../../base/browser/mouseEvent.js';
+import { ILanguageService } from '../../../../editor/common/languages/language.js';
+import { IModelService } from '../../../../editor/common/services/model.js';
+import { getFlatContextMenuActions } from '../../../../platform/actions/browser/menuEntryActionViewItem.js';
+import { IContextMenuService } from '../../../../platform/contextview/browser/contextView.js';
+import { ResourceContextKey } from '../../../common/contextkeys.js';
+import { Location, SymbolKind } from '../../../../editor/common/languages.js';
+
 
 abstract class AbstractChatAttachmentWidget extends Disposable {
 	public readonly element: HTMLElement;
@@ -49,7 +63,7 @@ abstract class AbstractChatAttachmentWidget extends Disposable {
 
 	constructor(
 		private readonly attachment: IChatRequestVariableEntry,
-		private readonly shouldFocusClearButton: boolean,
+		private readonly options: { shouldFocusClearButton: boolean; supportsDeletion: boolean },
 		container: HTMLElement,
 		contextResourceLabels: ResourceLabels,
 		protected readonly hoverDelegate: IHoverDelegate,
@@ -70,7 +84,7 @@ abstract class AbstractChatAttachmentWidget extends Disposable {
 
 	protected attachClearButton() {
 
-		if (this.attachment.range) {
+		if (this.attachment.range || !this.options.supportsDeletion) {
 			// no clear button for attachments with ranges because range means
 			// referenced from prompt
 			return;
@@ -91,7 +105,7 @@ abstract class AbstractChatAttachmentWidget extends Disposable {
 				this._onDidDelete.fire(e.browserEvent);
 			}
 		}));
-		if (this.shouldFocusClearButton) {
+		if (this.options.shouldFocusClearButton) {
 			clearButton.focus();
 		}
 	}
@@ -150,7 +164,7 @@ export class FileAttachmentWidget extends AbstractChatAttachmentWidget {
 		range: IRange | undefined,
 		attachment: IChatRequestVariableEntry,
 		currentLanguageModel: ILanguageModelChatMetadataAndIdentifier | undefined,
-		shouldFocusClearButton: boolean,
+		options: { shouldFocusClearButton: boolean; supportsDeletion: boolean },
 		container: HTMLElement,
 		contextResourceLabels: ResourceLabels,
 		hoverDelegate: IHoverDelegate,
@@ -161,7 +175,7 @@ export class FileAttachmentWidget extends AbstractChatAttachmentWidget {
 		@ILanguageModelsService private readonly languageModelsService: ILanguageModelsService,
 		@IInstantiationService private readonly instantiationService: IInstantiationService,
 	) {
-		super(attachment, shouldFocusClearButton, container, contextResourceLabels, hoverDelegate, currentLanguageModel, commandService, openerService);
+		super(attachment, options, container, contextResourceLabels, hoverDelegate, currentLanguageModel, commandService, openerService);
 
 		const fileBasename = basename(resource.path);
 		const fileDirname = dirname(resource.path);
@@ -215,7 +229,7 @@ export class ImageAttachmentWidget extends AbstractChatAttachmentWidget {
 		resource: URI | undefined,
 		attachment: IChatRequestVariableEntry,
 		currentLanguageModel: ILanguageModelChatMetadataAndIdentifier | undefined,
-		shouldFocusClearButton: boolean,
+		options: { shouldFocusClearButton: boolean; supportsDeletion: boolean },
 		container: HTMLElement,
 		contextResourceLabels: ResourceLabels,
 		hoverDelegate: IHoverDelegate,
@@ -225,7 +239,7 @@ export class ImageAttachmentWidget extends AbstractChatAttachmentWidget {
 		@ILanguageModelsService private readonly languageModelsService: ILanguageModelsService,
 		@ITelemetryService private readonly telemetryService: ITelemetryService,
 	) {
-		super(attachment, shouldFocusClearButton, container, contextResourceLabels, hoverDelegate, currentLanguageModel, commandService, openerService);
+		super(attachment, options, container, contextResourceLabels, hoverDelegate, currentLanguageModel, commandService, openerService);
 
 		let ariaLabel: string;
 		if (attachment.omittedState === OmittedState.Full) {
@@ -351,7 +365,7 @@ export class PasteAttachmentWidget extends AbstractChatAttachmentWidget {
 	constructor(
 		attachment: IChatRequestPasteVariableEntry,
 		currentLanguageModel: ILanguageModelChatMetadataAndIdentifier | undefined,
-		shouldFocusClearButton: boolean,
+		options: { shouldFocusClearButton: boolean; supportsDeletion: boolean },
 		container: HTMLElement,
 		contextResourceLabels: ResourceLabels,
 		hoverDelegate: IHoverDelegate,
@@ -360,7 +374,7 @@ export class PasteAttachmentWidget extends AbstractChatAttachmentWidget {
 		@IHoverService private readonly hoverService: IHoverService,
 		@IInstantiationService private readonly instantiationService: IInstantiationService,
 	) {
-		super(attachment, shouldFocusClearButton, container, contextResourceLabels, hoverDelegate, currentLanguageModel, commandService, openerService);
+		super(attachment, options, container, contextResourceLabels, hoverDelegate, currentLanguageModel, commandService, openerService);
 
 		const ariaLabel = localize('chat.attachment', "Attached context, {0}", attachment.name);
 		this.element.ariaLabel = ariaLabel;
@@ -406,7 +420,7 @@ export class DefaultChatAttachmentWidget extends AbstractChatAttachmentWidget {
 		range: IRange | undefined,
 		attachment: IChatRequestVariableEntry,
 		currentLanguageModel: ILanguageModelChatMetadataAndIdentifier | undefined,
-		shouldFocusClearButton: boolean,
+		options: { shouldFocusClearButton: boolean; supportsDeletion: boolean },
 		container: HTMLElement,
 		contextResourceLabels: ResourceLabels,
 		hoverDelegate: IHoverDelegate,
@@ -415,7 +429,7 @@ export class DefaultChatAttachmentWidget extends AbstractChatAttachmentWidget {
 		@IContextKeyService private readonly contextKeyService: IContextKeyService,
 		@IInstantiationService private readonly instantiationService: IInstantiationService,
 	) {
-		super(attachment, shouldFocusClearButton, container, contextResourceLabels, hoverDelegate, currentLanguageModel, commandService, openerService);
+		super(attachment, options, container, contextResourceLabels, hoverDelegate, currentLanguageModel, commandService, openerService);
 
 		const attachmentLabel = attachment.fullName ?? attachment.name;
 		const withIcon = attachment.icon?.id ? `$(${attachment.icon.id})\u00A0${attachmentLabel}` : attachmentLabel;
@@ -452,7 +466,7 @@ export class NotebookCellOutputChatAttachmentWidget extends AbstractChatAttachme
 		resource: URI,
 		attachment: INotebookOutputVariableEntry,
 		currentLanguageModel: ILanguageModelChatMetadataAndIdentifier | undefined,
-		shouldFocusClearButton: boolean,
+		options: { shouldFocusClearButton: boolean; supportsDeletion: boolean },
 		container: HTMLElement,
 		contextResourceLabels: ResourceLabels,
 		hoverDelegate: IHoverDelegate,
@@ -463,7 +477,7 @@ export class NotebookCellOutputChatAttachmentWidget extends AbstractChatAttachme
 		@INotebookService private readonly notebookService: INotebookService,
 		@IInstantiationService private readonly instantiationService: IInstantiationService,
 	) {
-		super(attachment, shouldFocusClearButton, container, contextResourceLabels, hoverDelegate, currentLanguageModel, commandService, openerService);
+		super(attachment, options, container, contextResourceLabels, hoverDelegate, currentLanguageModel, commandService, openerService);
 
 		switch (attachment.mimeType) {
 			case 'application/vnd.code.notebook.error': {
@@ -549,7 +563,7 @@ export class ElementChatAttachmentWidget extends AbstractChatAttachmentWidget {
 	constructor(
 		attachment: IElementVariableEntry,
 		currentLanguageModel: ILanguageModelChatMetadataAndIdentifier | undefined,
-		shouldFocusClearButton: boolean,
+		options: { shouldFocusClearButton: boolean; supportsDeletion: boolean },
 		container: HTMLElement,
 		contextResourceLabels: ResourceLabels,
 		hoverDelegate: IHoverDelegate,
@@ -557,7 +571,7 @@ export class ElementChatAttachmentWidget extends AbstractChatAttachmentWidget {
 		@IOpenerService openerService: IOpenerService,
 		@IEditorService editorService: IEditorService,
 	) {
-		super(attachment, shouldFocusClearButton, container, contextResourceLabels, hoverDelegate, currentLanguageModel, commandService, openerService);
+		super(attachment, options, container, contextResourceLabels, hoverDelegate, currentLanguageModel, commandService, openerService);
 
 		const ariaLabel = localize('chat.elementAttachment', "Attached element, {0}", attachment.name);
 		this.element.ariaLabel = ariaLabel;
@@ -581,4 +595,114 @@ export class ElementChatAttachmentWidget extends AbstractChatAttachmentWidget {
 
 		this.attachClearButton();
 	}
+}
+
+export function hookUpResourceAttachmentDragAndContextMenu(accessor: ServicesAccessor, widget: HTMLElement, resource: URI): IDisposable {
+	const contextKeyService = accessor.get(IContextKeyService);
+	const instantiationService = accessor.get(IInstantiationService);
+
+	const store = new DisposableStore();
+
+	// Context
+	const scopedContextKeyService = store.add(contextKeyService.createScoped(widget));
+	store.add(setResourceContext(accessor, scopedContextKeyService, resource));
+
+	// Drag and drop
+	widget.draggable = true;
+	store.add(dom.addDisposableListener(widget, 'dragstart', e => {
+		instantiationService.invokeFunction(accessor => fillEditorsDragData(accessor, [resource], e));
+		e.dataTransfer?.setDragImage(widget, 0, 0);
+	}));
+
+	// Context menu
+	store.add(addBasicContextMenu(accessor, widget, scopedContextKeyService, MenuId.ChatInputResourceAttachmentContext, resource));
+
+	return store;
+}
+
+export function hookUpSymbolAttachmentDragAndContextMenu(accessor: ServicesAccessor, widget: HTMLElement, scopedContextKeyService: IScopedContextKeyService, attachment: { name: string; value: Location; kind: SymbolKind }, contextMenuId: MenuId): IDisposable {
+	const instantiationService = accessor.get(IInstantiationService);
+	const languageFeaturesService = accessor.get(ILanguageFeaturesService);
+	const textModelService = accessor.get(ITextModelService);
+
+	const store = new DisposableStore();
+
+	// Context
+	store.add(setResourceContext(accessor, scopedContextKeyService, attachment.value.uri));
+
+	const chatResourceContext = chatAttachmentResourceContextKey.bindTo(scopedContextKeyService);
+	chatResourceContext.set(attachment.value.uri.toString());
+
+	// Drag and drop
+	widget.draggable = true;
+	store.add(dom.addDisposableListener(widget, 'dragstart', e => {
+		instantiationService.invokeFunction(accessor => fillEditorsDragData(accessor, [{ resource: attachment.value.uri, selection: attachment.value.range }], e));
+
+		fillInSymbolsDragData([{
+			fsPath: attachment.value.uri.fsPath,
+			range: attachment.value.range,
+			name: attachment.name,
+			kind: attachment.kind,
+		}], e);
+
+		e.dataTransfer?.setDragImage(widget, 0, 0);
+	}));
+
+	// Context menu
+	const providerContexts: ReadonlyArray<[IContextKey<boolean>, LanguageFeatureRegistry<unknown>]> = [
+		[EditorContextKeys.hasDefinitionProvider.bindTo(scopedContextKeyService), languageFeaturesService.definitionProvider],
+		[EditorContextKeys.hasReferenceProvider.bindTo(scopedContextKeyService), languageFeaturesService.referenceProvider],
+		[EditorContextKeys.hasImplementationProvider.bindTo(scopedContextKeyService), languageFeaturesService.implementationProvider],
+		[EditorContextKeys.hasTypeDefinitionProvider.bindTo(scopedContextKeyService), languageFeaturesService.typeDefinitionProvider],
+	];
+
+	const updateContextKeys = async () => {
+		const modelRef = await textModelService.createModelReference(attachment.value.uri);
+		try {
+			const model = modelRef.object.textEditorModel;
+			for (const [contextKey, registry] of providerContexts) {
+				contextKey.set(registry.has(model));
+			}
+		} finally {
+			modelRef.dispose();
+		}
+	};
+	store.add(addBasicContextMenu(accessor, widget, scopedContextKeyService, contextMenuId, attachment.value, updateContextKeys));
+
+	return store;
+}
+
+function setResourceContext(accessor: ServicesAccessor, scopedContextKeyService: IScopedContextKeyService, resource: URI) {
+	const fileService = accessor.get(IFileService);
+	const languageService = accessor.get(ILanguageService);
+	const modelService = accessor.get(IModelService);
+
+	const resourceContextKey = new ResourceContextKey(scopedContextKeyService, fileService, languageService, modelService);
+	resourceContextKey.set(resource);
+	return resourceContextKey;
+}
+
+function addBasicContextMenu(accessor: ServicesAccessor, widget: HTMLElement, scopedContextKeyService: IScopedContextKeyService, menuId: MenuId, arg: any, updateContextKeys?: () => Promise<void>): IDisposable {
+	const contextMenuService = accessor.get(IContextMenuService);
+	const menuService = accessor.get(IMenuService);
+
+	return dom.addDisposableListener(widget, dom.EventType.CONTEXT_MENU, async domEvent => {
+		const event = new StandardMouseEvent(dom.getWindow(domEvent), domEvent);
+		dom.EventHelper.stop(domEvent, true);
+
+		try {
+			await updateContextKeys?.();
+		} catch (e) {
+			console.error(e);
+		}
+
+		contextMenuService.showContextMenu({
+			contextKeyService: scopedContextKeyService,
+			getAnchor: () => event,
+			getActions: () => {
+				const menu = menuService.getMenuActions(menuId, scopedContextKeyService, { arg });
+				return getFlatContextMenuActions(menu);
+			},
+		});
+	});
 }
